refactor schedule into reusable
