+++ client-course-schedulizer/src/components/reuseables/Schedule/index.ts
@@ -1,0 +1,1 @@
+export * from "./Schedule";
